feat(results): add DOCX/PDF export to SimpleDeepAgent report
Extract the Reviewer 2 markdown export (DOCX via @mohtasham/md-to-docx,
PDF via print window) into a reusable MarkdownDownloadButton component and
reuse it for the SimpleDeepAgent "Report" output.

Also patch @mohtasham/md-to-docx so blockquotes preserve newlines: the
library put multi-line quote text into a single run, which Word collapses.
The patch emits one run per line with a break before each line after the
first.
