+++ frontend/package.json
@@ -97,5 +97,10 @@
     "tw-animate-css": "^1.3.8",
     "typescript": "^5",
     "unified": "^11.0.5"
+  },
+  "pnpm": {
+    "patchedDependencies": {
+      "@mohtasham/md-to-docx@2.9.0": "patches/@mohtasham__md-to-docx@2.9.0.patch"
+    }
   }
 }
